normalization to norm
