+++ src/CMakeLists.txt
@@ -131,7 +131,7 @@ set( MIOpen_Source
     logger.cpp
     layernorm_api.cpp
     lrn_api.cpp
-    normalization/problem_description.cpp
+    norm/problem_description.cpp
     op_args.cpp
     operator.cpp
     performance_config.cpp
@@ -167,8 +167,8 @@ set( MIOpen_Source
     solver/batchnorm/forward_spatial_multiple.cpp
     solver/batchnorm/forward_spatial_single.cpp
     solver/batchnorm/forward_training_ck.cpp
-    solver/normalization/forward_layernorm.cpp
-    solver/normalization/forward_layernorm2d_ck.cpp
+    solver/norm/forward_layernorm.cpp
+    solver/norm/forward_layernorm2d_ck.cpp
     solver/conv_asm_1x1u.cpp
     solver/conv_asm_1x1u_bias_activ_fused.cpp
     solver/conv_asm_1x1u_stride2.cpp
